+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 -   `Display.border_size` fixed and integrated into `Input.get_mouse_pos()`: you only get on-renderer positions.
 -   `Image.from_surface()` and `Image.from_buffer` functionality.
 -   `Group.active` property. Boolean that controls whether the group is drawn and updated.
+-   `Color.random_default()` Allows to randomly choose from a set of default colors.
 
 ### Changed